[feature/#50] move large and shared components to the `components/` folder
