+++ src/components/BrandsManager/BrandsManager.module.css
@@ -1,4 +1,4 @@
-@reference '../../../globals.css';
+@reference '../../app/globals.css';
 
 .editable-brand-tile {
 	@apply flex flex-row gap-2 items-center;
