+++ src/components/BrandsManager/BrandsManager.tsx
